mkbgp: Optionally set unreachable peers as passive
By passing -P <timeout>, the script now checks if the BGP peer is
reachable by opening a TCP connection to the BGP port. If it isn't, it
is added to the config nonetheless but only as passive peer, meaning
that incoming connections from this peer will be accepted but no
attempts to establish a connection is made.

NOTE: Only bird supports the passive peer feature, so only bird is
supported by this new option. While supplying -P together with
"-f quagga" isn't an error and will still do the checking, it doesn't
change the resulting config in any way.